Enhance assistant functionality and testing framework
- Updated `AssistantEndpointTests.cs` with new test methods for complex assistant flows, simple thread flows, and file uploads.
- Modified `Rystem.OpenAi.UnitTests.csproj` to include new text files for testing.
- Changed `scoreThreshold` type from `int` to `float` in `IOpenAiFileSearchAssistant.cs`.
- Replaced `UseFileSearch` with a generic `Use` method in `IOpenAiFileSearchToolResourcesAssistant.cs`.
- Added `WithFileSearch` and `WithVectorStoresAsFileSearch` methods in `IOpenAiToolResourcesAssistant.cs`.
- Updated property types in various classes to allow for more flexible management.
- Changed `ScoreThreshold` type to `float?` in `AssistantRankerSettingsForFileSearchTool.cs`.
- Rearranged `Tools` property types in `AssistantRequest.cs`.
- Enhanced metadata handling and file management in `OpenAiFileSearchAssistant.cs` and `OpenAiFileSearchToolResourcesAssistant.cs`.
- Added a new `Response` property in `HttpClientExtensions.cs` for improved error handling.
- Introduced narrative content in `chapter01.txt` and `chapter02.txt` to enrich the assistant's context.

+++ src/Rystem.OpenAi/Endpoints/Assistant/Assistants/OpenAiFileSearchAssistant.cs
@@ -10,7 +10,7 @@
             _builder = builder;
             _assistantFileSearchTool = assistantFileSearchTool;
         }
-        public T UseCustom(RankerName ranker, int scoreThreshold)
+        public T UseCustom(RankerName ranker, float scoreThreshold)
         {
             _assistantFileSearchTool.FileSearch ??= new();
             _assistantFileSearchTool.FileSearch.RankingOptions ??= new();
